moves a few of the utils out of builder, especially those use by traverse
