+++ packages/runner/src/builder/opaque-ref.ts
@@ -14,7 +14,7 @@ import {
   type UnsafeBinding,
 } from "./types.ts";
 import { ContextualFlowControl } from "../cfc.ts";
-import { hasValueAtPath, setValueAtPath } from "./path-utils.ts";
+import { hasValueAtPath, setValueAtPath } from "../path-utils.ts";
 import { getTopFrame, recipe } from "./recipe.ts";
 import { createNodeFactory } from "./module.ts";
