+++ packages/runner/src/builder/recipe.ts
@@ -32,8 +32,8 @@ import {
   recipeToJSON,
   toJSONWithAliases,
 } from "./json-utils.ts";
-import { setValueAtPath } from "./path-utils.ts";
-import { traverseValue } from "./traverse-utils.ts";
+import { setValueAtPath } from "../path-utils.ts";
+import { traverseValue } from "../traverse-utils.ts";
 
 /** Declare a recipe
  *
