+++ packages/runner/src/path-utils.ts
@@ -1,5 +1,4 @@
 import { isRecord } from "@commontools/utils/types";
-import { deepEqual } from "./traverse-utils.ts";
 
 export function setValueAtPath(
   obj: any,
@@ -49,3 +48,19 @@ export function hasValueAtPath(obj: any, path: PropertyKey[]): boolean {
   }
   return current !== undefined;
 }
+
+export const deepEqual = (a: any, b: any): boolean => {
+  if (a === b) return true;
+  if (isRecord(a) && isRecord(b)) {
+    if (a.constructor !== b.constructor) return false;
+    const keysA = Object.keys(a);
+    const keysB = Object.keys(b);
+    if (keysA.length !== keysB.length) return false;
+    for (const key of keysA) {
+      if (!keysB.includes(key)) return false;
+      if (!deepEqual(a[key], b[key])) return false;
+    }
+    return true;
+  }
+  return a !== a && b !== b; // NaN check
+};
