+++ packages/runner/src/storage/cache.ts
@@ -1,9 +1,9 @@
 import { fromString, refer } from "merkle-reference";
 import { isBrowser } from "@commontools/utils/env";
 import { isObject } from "@commontools/utils/types";
-import { JSONSchema } from "../builder/types.ts";
+import { type JSONSchema } from "../builder/types.ts";
 import { ContextualFlowControl } from "../cfc.ts";
-import { deepEqual } from "../builder/traverse-utils.ts";
+import { deepEqual } from "../path-utils.ts";
 import { MapSet } from "../traverse.ts";
 import type {
   AuthorizationError,
